requirements.txt
requests>=2.28
python-dotenv>=1.0
loguru>=0.7.0
sqlalchemy>=1.4
alembic>=1.11
pytest>=7.0
websocket-client>=1.6
pydantic>=1.10
psycopg2-binary>=2.9
pandas>=1.5
numpy>=1.23